Extract Browser/WebView specific codes from lifecycle code
