+++ common/src/main/java/com/microsoft/identity/common/internal/providers/oauth2/AuthorizationActivity.java
@@ -4,15 +4,24 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentTransaction;
 
 import com.microsoft.identity.common.R;
 import com.microsoft.identity.common.internal.logging.DiagnosticContext;
+import com.microsoft.identity.common.internal.telemetry.Telemetry;
+import com.microsoft.identity.common.internal.telemetry.events.UiStartEvent;
 import com.microsoft.identity.common.internal.ui.AuthorizationAgent;
 
 import java.util.HashMap;
+
+import static com.microsoft.identity.common.adal.internal.AuthenticationConstants.AuthorizationIntentParameters.KEY_AUTH_AUTHORIZATION_AGENT;
+import static com.microsoft.identity.common.adal.internal.AuthenticationConstants.AuthorizationIntentParameters.KEY_AUTH_INTENT;
+import static com.microsoft.identity.common.adal.internal.AuthenticationConstants.AuthorizationIntentParameters.KEY_AUTH_REDIRECT_URI;
+import static com.microsoft.identity.common.adal.internal.AuthenticationConstants.AuthorizationIntentParameters.KEY_AUTH_REQUEST_URL;
+import static com.microsoft.identity.common.adal.internal.AuthenticationConstants.AuthorizationIntentParameters.KEY_REQUEST_HEADERS;
 
 public final class AuthorizationActivity extends FragmentActivity {
 
@@ -25,22 +34,36 @@ public final class AuthorizationActivity extends FragmentActivity {
                                            final HashMap<String, String> requestHeaders,
                                            final AuthorizationAgent authorizationAgent) {
         final Intent intent = new Intent(context, AuthorizationActivity.class);
-        intent.putExtra(AuthorizationFragment.KEY_AUTH_INTENT, authIntent);
-        intent.putExtra(AuthorizationFragment.KEY_AUTH_REQUEST_URL, requestUrl);
-        intent.putExtra(AuthorizationFragment.KEY_AUTH_REDIRECT_URI, redirectUri);
-        intent.putExtra(AuthorizationFragment.KEY_REQUEST_HEADERS, requestHeaders);
-        intent.putExtra(AuthorizationFragment.KEY_AUTH_AUTHORIZATION_AGENT, authorizationAgent);
+        intent.putExtra(KEY_AUTH_INTENT, authIntent);
+        intent.putExtra(KEY_AUTH_REQUEST_URL, requestUrl);
+        intent.putExtra(KEY_AUTH_REDIRECT_URI, redirectUri);
+        intent.putExtra(KEY_REQUEST_HEADERS, requestHeaders);
+        intent.putExtra(KEY_AUTH_AUTHORIZATION_AGENT, authorizationAgent);
         intent.putExtra(DiagnosticContext.CORRELATION_ID, DiagnosticContext.getRequestContext().get(DiagnosticContext.CORRELATION_ID));
         return intent;
     }
+
+    public static AuthorizationFragment getAuthorizationFragmentFromStartIntent(@NonNull final Intent intent){
+        AuthorizationFragment fragment;
+        final AuthorizationAgent authorizationAgent = (AuthorizationAgent)intent.getSerializableExtra(KEY_AUTH_AUTHORIZATION_AGENT);
+        Telemetry.emit(new UiStartEvent().putUserAgent(authorizationAgent));
+
+        if (authorizationAgent == AuthorizationAgent.WEBVIEW) {
+            fragment = new WebViewAuthorizationFragment();
+        } else {
+            fragment = new BrowserAuthorizationFragment();
+        }
+
+        fragment.setInstanceState(intent.getExtras());
+        return fragment;
+    }
     
     @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
+    public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.authorization_activity);
-        mFragment = new AuthorizationFragment();
-        mFragment.setInstanceState(getIntent().getExtras());
 
+        mFragment = getAuthorizationFragmentFromStartIntent(getIntent());
         getSupportFragmentManager()
                 .beginTransaction()
                 .setTransitionStyle(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
